#Stage 3 passed

+++ application/libraries/VpxMigration.php
@@ -32,7 +32,6 @@ class VpxMigration {
     /*
      * defaults;
      */
-
     function __construct($params = null) {
         // parent::__construct();
         isset($this->ci) OR $this->ci = get_instance();
@@ -42,8 +41,8 @@ class VpxMigration {
         $this->db_host = $this->ci->db_master->hostname;
         $this->db_name = $this->ci->db_master->database;
         $this->path = APPPATH . $this->path;
-        if ($params)
-            $this->init_config($params);
+        
+        if ($params) $this->init_config($params);
     }
 
     /**
@@ -71,31 +70,25 @@ class VpxMigration {
      * @param string $tables
      * @return boolean|string
      */
-    function generate($tables = null, $dbs = null) {
-        if(!$dbs)
-            $dbs = 'sample_db';
-        if ($tables)
-            $this->tables = $tables;
+    function generate($tables = null) {
+        $return = $up = $down = '';
 
-        $return = '';
+        if ($tables) { $this->tables = $tables; }
+
         /* open file */
-        if ($this->write_file)
-        {
-            if (!is_dir($this->path) OR !is_really_writable($this->path))
-            {
+        if ($this->write_file) {
+            if (!is_dir($this->path) OR !is_really_writable($this->path)) {
                 $msg = "Unable to write migration file: " . $this->path;
                 log_message('error', $msg);
                 echo $msg;
                 return;
             }
 
-            if (!$this->file_per_table)
-            {
-                $file_path = $this->path . '/' . $this->file_name . '.sql';
+            if (!$this->file_per_table) {
+                $file_path = $this->path.'/'.$this->file_name.'.sql';
                 $file = fopen($file_path, 'w+');
 
-                if (!$file)
-                {
+                if (!$file) {
                     $msg = "no file";
                     log_message('error', $msg);
                     echo $msg;
@@ -104,11 +97,9 @@ class VpxMigration {
             }
         }
 
-
         // if default, then run all tables, otherwise just do the list provided
-        if ($this->tables == '*'){
-
-            $query = $this->ci->db_master->query('SHOW full TABLES FROM ' . $this->ci->db_master->protect_identifiers($this->ci->db_master->database));
+        if ($this->tables == '*') {
+            $query = $this->ci->db_master->query('SHOW full TABLES FROM '.$this->ci->db_master->protect_identifiers($this->ci->db_master->database));
 
             $retval = array();
 
@@ -137,14 +128,30 @@ class VpxMigration {
 
             $this->tables = array();
             $this->tables = $retval;
-        } else {
+        }
+        else {
+            if(!is_array($tables) && !$this->ci->db_master->table_exists($tables)) {
+                echo 'Table Not Found!';
+                exit;
+            }
+            
+            if(is_array($tables)) {
+                foreach ($tables as $table) {
+                    if(!$this->ci->db_master->table_exists($table)) {
+                        if (($key = array_search($table, $tables)) !== false) {
+                            unset($tables[$key]);
+                        }
+                    }
+                }
+            }
+            
             $this->tables = is_array($tables) ? $tables : explode(',', $tables);
         }
 
         ## if write file, check if we can
         if ($this->write_file) {
             /* make subdir */
-            $path = $this->path . '/' . $this->file_name;
+            $path = $this->path.'/'.$this->file_name;
 
             if (!@is_dir($path)) {
                 if (!@mkdir($path, DIR_WRITE_MODE, true)) {
@@ -161,9 +168,9 @@ class VpxMigration {
             }
 
             if($tables)
-                $file_path = $path . '/' . date('YmdHis') . '_create_' . $tables . '.php';
+                $file_path = $path.'/'.date('YmdHis').'_create_'.$tables.'.php';
             else
-                $file_path = $path . '/'.date('YmdHis').'_create_base.php';
+                $file_path = $path.'/'.date('YmdHis').'_create_base.php';
             
             $file = fopen($file_path, 'w+');
 
@@ -176,63 +183,60 @@ class VpxMigration {
             }
         }
 
-
-        $up = '';
-        $down = '';
         //loop through tables
         foreach ($this->tables as $table) {
-            log_message('debug', print_r($table, true));
+            // log_message('debug', print_r($table, true));
 
-            $q = $this->ci->db_master->query('describe ' . $this->ci->db_master->protect_identifiers($this->ci->db_master->database . '.' . $table));
+            if($table) {
+                $q = $this->ci->db_master->query('describe '.$this->ci->db_master->protect_identifiers($this->ci->db_master->database.'.'.$table));
             
-            // No result means the table name was invalid
-            if ($q === FALSE) {
-                continue;
+                // No result means the table name was invalid
+                if ($q === FALSE) { continue; }
+
+                $columns = $q->result_array();
+
+                $q = $this->ci->db_master->query(' SHOW TABLE STATUS WHERE Name = \'' . $table . '\'');
+
+                $engines = $q->row_array();
+
+                $up .= "\n\t" . 'if($this->db_exists()) {';
+                $up .= "\n\t\t" . '## Create Table ' . $table . "\n";
+                foreach ($columns as $column)
+                {
+                    $up .= "\t\t" . '$this->dbforge->add_field("' . "`$column[Field]` $column[Type] " . ($column['Null'] == 'NO' ? 'NOT NULL' : 'NULL') .
+                            (
+                            #  if its timestamp column, don't '' around default value .... crap way, but should work for now
+                            $column['Default'] ? ' DEFAULT ' . ($column['Type'] == 'timestamp' ? $column['Default'] : '\'' . $column['Default'] . '\'') : ''
+                            )
+                            . " $column[Extra]\");" . "\n";
+
+                    if ($column['Key'] == 'PRI')
+                        $up .= "\t\t" . '$this->dbforge->add_key("' . $column['Field'] . '",true);' . "\n";
+                }
+                $up .= "\t\t" . '$this->dbforge->create_table("' . $table . '", TRUE);' . "\n";
+                if (isset($engines['Engine']) and $engines['Engine'])
+                    $up .= "\t\t" . '$this->db->query(\'ALTER TABLE  ' . $this->ci->db_master->protect_identifiers($table) . ' ENGINE = ' . $engines['Engine']. '\');'."\n\t". '}';
+
+
+                $down .= "\t\t" . '## Drop table ' . $table . ' ##' . "\n";
+                $down .= "\t\t" . '$this->dbforge->drop_table("' . $table . '", TRUE);' . "\n";
+
+                /* clear some mem */
+                $q->free_result();
             }
-
-            $columns = $q->result_array();
-
-            $q = $this->ci->db_master->query(' SHOW TABLE STATUS WHERE Name = \'' . $table . '\'');
-            $engines = $q->row_array();
-
-
-            $up .= "\n\t" . 'if($this->db_exists()) {';
-            $up .= "\n\t\t" . '## Create Table ' . $table . "\n";
-            foreach ($columns as $column)
-            {
-                $up .= "\t\t" . '$this->dbforge->add_field("' . "`$column[Field]` $column[Type] " . ($column['Null'] == 'NO' ? 'NOT NULL' : 'NULL') .
-                        (
-                        #  if its timestamp column, don't '' around default value .... crap way, but should work for now
-                        $column['Default'] ? ' DEFAULT ' . ($column['Type'] == 'timestamp' ? $column['Default'] : '\'' . $column['Default'] . '\'') : ''
-                        )
-                        . " $column[Extra]\");" . "\n";
-
-                if ($column['Key'] == 'PRI')
-                    $up .= "\t\t" . '$this->dbforge->add_key("' . $column['Field'] . '",true);' . "\n";
-            }
-            $up .= "\t\t" . '$this->dbforge->create_table("' . $table . '", TRUE);' . "\n";
-            if (isset($engines['Engine']) and $engines['Engine'])
-                $up .= "\t\t" . '$this->db->query(\'ALTER TABLE  ' . $this->ci->db_master->protect_identifiers($table) . ' ENGINE = ' . $engines['Engine']. '\');'."\n\t". '}';
-
-
-            $down .= "\t\t" . '### Drop table ' . $table . ' ##' . "\n";
-            $down .= "\t\t" . '$this->dbforge->drop_table("' . $table . '", TRUE);' . "\n";
-
-            /* clear some mem */
-            $q->free_result();
         }
-
+        
         ### generate the text ##
         $return .= '<?php ';
         $return .= 'defined(\'BASEPATH\') OR exit(\'No direct script access allowed\');' . "\n\n";
         $return .= 'class Migration_create_base extends CI_Migration {' . "\n";
 
-        $return .= "\n\t" . 'public function db_exists {
+        $return .= "\n\t".'public function db_exists {
             $this->load->dbutil();
             $this->load->dbforge();
 
-            if (!$this->dbutil->database_exists(\''.$dbs.'\')) {
-                if ($this->dbforge->create_database(\''.$dbs.'\')) {
+            if (!$this->dbutil->database_exists(\''.$this->db_name.'\')) {
+                if ($this->dbforge->create_database(\''.$this->db_name.'\')) {
                     return TRUE;
                 }
                 else {
@@ -242,26 +246,22 @@ class VpxMigration {
         }' . "\n";
 
         $return .= "\n\t" . 'public function up() {' . "\n";
-
         $return .= $up;
-        $return .= "\n\t" . ' }' . "\n";
+        $return .= "\n\t}\n";
 
-        $return .= "\n\t" . 'public function down()';
-        $return .= "\t" . '{' . "\n";
-        $return .= $down . "\n";
-        $return .= "\t" . '}' . "\n" . '}';
+        $return .= "\n\t"."public function down() {\n\t\t";
+        $return .= $down;
+        $return .= "\n\t}\n}";
 
         ## write the file, or simply return if write_file false
-        if ($this->write_file)
-        {
+        if ($this->write_file) {
             fwrite($file, $return);
             fclose($file);
-            echo "Create file migration with success!";
-            return true;
-        } else
-        {
-            return $return;
+            echo "Migration file/s Created  successfully!";
+            exit;
+            //return true;
         }
+        else { return $return; }
     }
 
 }
